backend/requirements.txt
Flask
Flask-Cors
vapi-server-sdk
google-generativeai
python-dotenv
Pillow 